Only release claim on certain calls to DESTROY
When destruction happens, it may be in the context of another process or thread which was forked from the original process/thread that made the claim.

This change keeps track of the original process (and thread if applicable) so that release only happens for the original process which created the claim.